backend/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-environ==0.11.2
gunicorn==22.0.0
Markdown==3.2.1
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.4.2
django-cors-headers
django-csp